streaming_app/requirements_streaming.txt
# DyStream Streaming Server Requirements

# FastAPI and WebSocket
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# All existing dependencies from main requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
transformers>=4.35.0
diffusers>=0.24.0
PyYAML>=6.0
scipy>=1.11.0
tqdm>=4.66.0
einops>=0.7.0
omegaconf>=2.3.0
pytorch-lightning>=2.1.0
wandb>=0.16.0
imageio>=2.37.0
imageio-ffmpeg>=0.6.0
moviepy>=1.0.3
librosa>=0.11.0
soundfile>=0.13.0

# Additional dependencies for streaming
pydantic>=2.0.0
aiofiles>=23.2.1
